Incorporating ActiveSupport and Sass
This project now uses ActiveSupport for help with text formatting.
Specifically we are using the titleize method, which is better than the
one I wrote.

We are also "using" sass.  I don't plan for it to be a dependency for this
project, but I am using it to help generate our stylesheet.  I need some
serious styling help.