Moving "matchers" / "ignore patterns" into Extractor class methods.
Matchers and ignore patterns are now in a central location under the
Extractor class.  This should make it easy to break out extractors out
into seperate files and have them register the appropriate matchers from
there.

This should also pave the way for allowing the user to specify custom
matches and ignores.