Adding SnoOut to ignore.  Fixing ignore bug.
SnoOut should be ignored by default since if a output directory is not
specified the output is nested under the input in a directory called SnoOut.

Also we need to chomp the lines we read from the ignore file.  Were not gonna
match when lines have "\n" at the end of them.