Turning this project into a proper rubygem
This project is now a proper rubygem.  I used bundle's style of
creating this directory.  Since this was an existing project I created
a test project using the command "bundle gem foo".  I think copied
the structure into this project.

Using bundler will make it easier for others to clone the repo
and work on this project in the future.  It also helps me to ensure
that all my dependencies are listed.