Adding breadcrumbs and removing the file name header on each page.
Breadcrumbs have been added and the titleized filename no longer
displays as a header on each page.  The code however is in bad
shape and I think its time to refactor before adding new features.

Lazy style behavior is being done in dangerous places.  See the link
method of the IndexPage for example.

Also we really need to split extractors up or something.