Getting rid of markdown intra emphasis
Its really annoying that mark down considers things
between a pair of '_' to be emphasis.  I have removed
this to make sno more programmer friendly.

You can still use emphasis by placing the text between
a pair of '*'.  This works fine and does not get in the
way of programmer style 'words'