Fixing a bug causing "display" on type ahead to be to long
When calling Sno with absolute paths the display value for type ahead
was way to long.  Previous we were just chomping off 2 levels of directory.
When called with an absoltue path, it may be much more than that.

To solve this we exposed another sudo global variable called OUTPUT_ROOT
so that we can properly regex off the right amount.  This is another ugly
hack that needs to be cleaned up in our first major refactor.