Adding syntax highlighting support
We now have syntax highlighting for markdown using coderay.
We are currently only doing syntax highlighting for markdown
because it is efficient to do so.

I will look to add this feature for other extractors soon, but will
want to come up with a way to only run it over a subset do to inefficiency.