Fixing resource copying
Previous resouces would never be copied without the -f options.
This fixes that, but we really need a better way of checking
whether or not we should copy.  We should probably use a generated
config file for this to see if we have the appropriate versioned
resources copied.

Fix for issue #2