Sno can now be hosted on a webserver using the "w" option
The implementation is nasty and needs refactoring but it works.
You can now specify a "w" option, see help for more information.
This option builds sno so that is hostable on a webserver instead
of a file system.  You can also provide a optional web root so that
it can be nested on the web server.

This should make sno projects easier to share across the web.